\name{albrand-package}
\alias{albrand-package}
\alias{albrand}
\docType{package}
\title{
  Antoine Lucas' Brand for Quarto Documents and Shiny Apps
}
\description{
  Bundles the 'al-brand' Quarto brand extension so any R workflow can
  produce branded HTML reports and themed Shiny apps from one identity.
}
\seealso{
  \code{\link[=render_report]{render_report()}},
  \code{\link[=use_albrand]{use_albrand()}},
  \code{\link[=bs_theme_brand]{bs_theme_brand()}}
}
